requirements.txt
Django==2.0.2
django-filter==1.1.0
djangorestframework==3.7.7
Markdown==2.6.11
pytz==2018.3
django-phonenumber-field==2.0.0
django-localflavor==2.0
requests==2.18.4
dj-database-url==0.5.0
psycopg2-binary==2.7.4